refactor: moved uploadedToCloudinary to separate helper
