+++ optimization/uploadToCloudinary.js
@@ -1,0 +1,16 @@
+const cloudinary = require("../library/cloudinary");
+
+const uploadToCloudinary = async (filePath, publicId) => {
+  try {
+    const result = await cloudinary.uploader.upload(filePath, {
+      public_id: publicId,
+      overwrite: true,
+    });
+    return result;
+  } catch (error) {
+    console.error(`Error uploading ${filePath} to Cloudinary:`, error);
+    return null;
+  }
+};
+
+module.exports = uploadToCloudinary;
